Separate the NewConfigurationProfile as a component. Prepare the import as an option using EllipsisOptions component
Now the user will have two ways of creating a Configuration Profile:
- As in the beginning, by typing the data in the edit form.
- The new option is by importing a JSON file as the structure (not yet designed)
